add payment and supply to the system in Market.Buy

+++ market14B/src/main/java/DomainLayer/backend/API/PaymentExternalService/PaymentBridge.java
@@ -4,7 +4,7 @@ public interface PaymentBridge {
 
     String handshake();
 
-    int pay(int amount,String currency,String card_number,int month,int year,String holder,String ccv);
+    int pay(double amount,String currency,String card_number,int month,int year,String holder,String ccv);
 
     int cancel_pay(int transaction_id);
